backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.21.1
SQLAlchemy>=2.0,<2.2
psycopg2-binary>=2.9
alembic==1.11.1
pydantic==1.10.11
python-jose==3.3.0
passlib[bcrypt]>=1.7.4,<1.8
python-dotenv==1.0.0
